LPS-187496 Decouple destination registration from ExportImportLifecycleManagerImpl to be able to remove service property and let it be implied as immediate.
