+++ modules/apps/export-import/export-import-service/src/main/java/com/liferay/exportimport/lifecycle/ExportImportLifecycleManagerImpl.java
@@ -17,25 +17,13 @@ package com.liferay.exportimport.lifecycle;
 import com.liferay.exportimport.kernel.lifecycle.ExportImportLifecycleEvent;
 import com.liferay.exportimport.kernel.lifecycle.ExportImportLifecycleEventFactory;
 import com.liferay.exportimport.kernel.lifecycle.ExportImportLifecycleManager;
-import com.liferay.portal.kernel.messaging.Destination;
-import com.liferay.portal.kernel.messaging.DestinationConfiguration;
-import com.liferay.portal.kernel.messaging.DestinationFactory;
 import com.liferay.portal.kernel.messaging.DestinationNames;
 import com.liferay.portal.kernel.messaging.Message;
 import com.liferay.portal.kernel.messaging.MessageBus;
-import com.liferay.portal.kernel.util.HashMapDictionaryBuilder;
 
 import java.io.Serializable;
 
-import java.util.Dictionary;
-import java.util.HashSet;
-import java.util.Set;
-
-import org.osgi.framework.BundleContext;
-import org.osgi.framework.ServiceRegistration;
-import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.service.component.annotations.Reference;
 
 /**
@@ -57,35 +45,6 @@ public class ExportImportLifecycleManagerImpl
 		fireExportImportLifecycleEvent(exportImportLifecycleEvent);
 	}
 
-	@Activate
-	protected void activate(BundleContext bundleContext) {
-		_bundleContext = bundleContext;
-
-		registerDestination(
-			bundleContext, DestinationConfiguration.DESTINATION_TYPE_SERIAL,
-			DestinationNames.EXPORT_IMPORT_LIFECYCLE_EVENT_ASYNC);
-		registerDestination(
-			bundleContext,
-			DestinationConfiguration.DESTINATION_TYPE_SYNCHRONOUS,
-			DestinationNames.EXPORT_IMPORT_LIFECYCLE_EVENT_SYNC);
-	}
-
-	@Deactivate
-	protected void deactivate() {
-		for (ServiceRegistration<Destination> serviceRegistration :
-				_serviceRegistrations) {
-
-			Destination destination = _bundleContext.getService(
-				serviceRegistration.getReference());
-
-			serviceRegistration.unregister();
-
-			destination.destroy();
-		}
-
-		_bundleContext = null;
-	}
-
 	protected void fireExportImportLifecycleEvent(
 		ExportImportLifecycleEvent exportImportLifecycleEvent) {
 
@@ -101,43 +60,11 @@ public class ExportImportLifecycleManagerImpl
 			message.clone());
 	}
 
-	protected ServiceRegistration<Destination> registerDestination(
-		BundleContext bundleContext, String destinationType,
-		String destinationName) {
-
-		DestinationConfiguration destinationConfiguration =
-			new DestinationConfiguration(destinationType, destinationName);
-
-		Destination destination = _destinationFactory.createDestination(
-			destinationConfiguration);
-
-		Dictionary<String, Object> dictionary =
-			HashMapDictionaryBuilder.<String, Object>put(
-				"destination.name", destination.getName()
-			).build();
-
-		ServiceRegistration<Destination> serviceRegistration =
-			bundleContext.registerService(
-				Destination.class, destination, dictionary);
-
-		_serviceRegistrations.add(serviceRegistration);
-
-		return serviceRegistration;
-	}
-
-	private volatile BundleContext _bundleContext;
-
-	@Reference
-	private DestinationFactory _destinationFactory;
-
 	@Reference
 	private ExportImportLifecycleEventFactory
 		_exportImportLifecycleEventFactory;
 
 	@Reference
 	private MessageBus _messageBus;
-
-	private final Set<ServiceRegistration<Destination>> _serviceRegistrations =
-		new HashSet<>();
 
 }
